pyproject.toml
[tool.poetry]
name = "semantic-model-generator"
version = "1.0.0"
description = "Curate a Semantic Model for ClickZetta Lakehouse"
authors = ["qililiang <qililiang@clickzetta.com>"]
license = "Apache Software License; BSD License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
pandas = "^2.0.1"
loguru = "^0.7.2"
clickzetta-connector-python = "0.8.92"
protobuf = "5.26.1"
pydantic = "2.8.2"
PyYAML = "^6.0.1"
"ruamel.yaml" = "0.17.21"
tqdm = "^4.66.5"
pyarrow = "14.0.2"
sqlglot = "25.10.0"
strictyaml = "^1.7.3"
streamlit = "1.36.0"
streamlit-extras = "0.4.0"
numpy = "^1.26.4"
python-dotenv = "^1.0.1"
urllib3 = "^1.26.19"
requests = "^2.32.3"
clickzetta-zettapark-python = "0.1.3"
dashscope = "^1.22.2"

# Optional dependencies for functionality such as partner semantic model support.
looker-sdk = { version = "^24.14.0", optional = true }

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^8.1.1"
types-pyyaml = "^6.0.12.20240311"
types-protobuf = "^4.24.0.20240311"
pip-licenses = "^4.4.0"
grpcio-tools = "1.64.1"

[tool.poetry.extras]
looker = ["looker-sdk"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
